src/teleop/scripts/requirements.txt
evdev==0.8.1
pyyaml